Update the endpoint flow paths
The new flow starts at the handler does a lookup req to flattener,
reponds to resource resolver which looks up the locations resource (and
eventually the practitioner) then passes on the data to the Bundler
which does lookup requests to create the encounter, observations and
composition resources. The Bundle is currently returned to the client
but eventually this should be sent to the HAPI-FHIR server in a response
that closes off the endpoint chain.

DSC19-25