Create and populate the file-queue volume
This creates an external volume which is prepopulated with the file queue config. This volume will then be used by the file queue. This is being done this way to enable the package to be run using the intantHIE command which uses a docker container to run other containers.

DSC19-114